predictive-smart-irrigation/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
ortools>=9.8.0
matplotlib>=3.7.0
plotly>=5.17.0
pytest>=7.4.0
requests>=2.31.0

